Flattened Java Dirs and Other Languages
Tomorrow, I'll be using the Java parts of this in jmage-processing.
I still need to fix throwing in lambdas, if that can even be done.
I guess I'll see tomorrow.

Option#consume2 would be nice or some way to branch without having
to use an if/else outside of the library.