Throwing in Lambda + None Singleton + .val_ptr();
Throwing in the lambda works but I needed to make a custom
functional interface.

None's constructor is now hidden and there is a single instance
that gets passed around. Option.None() does the dirty work of
casting it to satisfy the type system. I think it erases to the
same thing anyways.

Since None is a Singleton, you can check for pointer equality using
== between objects. I added a .val_ptr() which might come in handy.
Even if it doesn't, I still think that it's cool.