New Panic and Result::ignore
Panic::Panic is better than new Panic(msg).run() since it's less
verbose. Creating an Object seems kinda silly when you're trying to
panic. If it was throw new Panic(msg), that might make more sense.

With Result::ignore, it's just a nice way to ignore a potential
error. I'll be using this a lot with Thread::sleep.